hide reserved region within RandomAccessStore
